requirements/base.txt
argon2-cffi
whitenoise
pandas
psycopg2-binary

# Django
# ------------------------------------------------------------------------------
django
django-environ
django-jazzmin
# Django REST Framework
djangorestframework
django-cors-headers
djangorestframework-simplejwt
djoser
django-filter
